Products can now be created using API. Now all requests goes through our API

+++ Webshop/Webshop/Services/ProductImage.cs
@@ -107,6 +107,9 @@ namespace Webshop.Services
 
         public bool DeleteImage(string pathToImage)
         {
+            if (pathToImage == "")
+                return true;
+
             var fullPath = Path.Combine(_categoryFolderPath, pathToImage);
             File.Delete(fullPath);
             return ImageExist(fullPath);
